review(state): nonce + balance type fixes; restore DifficultyBombDelaysConverter
- AccountProof.Nonce: UInt256 → ulong (protocol nonce is ulong);
  ProofJsonConverter pulls the ulong converter for the field.
- BlockAccessListBasedWorldState.TryGetAccount: balance stays UInt256 instead
  of being truncated to ulong; AccountStruct ctor already takes UInt256.
- WorldState.IsNonZeroAccount: simplify !(Nonce == 0) → Nonce != 0.
- IWorldState.CommitTree gains int overload so callers writing
  for-loop counters don't need (ulong) casts (used by WorldStateManagerTests).
- WorldStateManagerTests.lastBlock: int → ulong; loop counter ulong;
  drops the (ulong)i / (ulong)lastBlock casts.
- BlockchainTestStreamingTracerTests: drop (ulong)(b + 1) cast (new int WithNumber).
- DifficultyBombDelaysConverter: restored — still referenced from
  EthashChainSpecEngineParameters.cs. (Deletion in the previous commit was a
  false positive: grep missed the [JsonConverter(typeof(...))] attribute usage.)

+++ src/Nethermind/Nethermind.State.Test.Runner.Test/BlockchainTestStreamingTracerTests.cs
@@ -45,7 +45,7 @@ public class BlockchainTestStreamingTracerTests
 
         for (int b = 0; b < blockCount; b++)
         {
-            tracer.StartNewBlockTrace(Build.A.Block.WithNumber((ulong)(b + 1)).TestObject);
+            tracer.StartNewBlockTrace(Build.A.Block.WithNumber(b + 1).TestObject);
             for (uint t = 0; t < txPerBlock; t++)
             {
                 tracer.StartNewTxTrace(Build.A.Transaction.WithValue(t + 1).WithNonce(t).TestObject);
